FIXED: Warnings about comparisons of signed and unsigned integers
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1487 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/doc/PdfPagesTree.cpp
@@ -583,7 +583,7 @@ void PdfPagesTree::InsertPagesIntoNode( PdfObject* pParent, const PdfObjectList 
     }
 
     // If new kids are still not pushed in then they may be appending to the end
-    if ( !bIsPushedIn && ( (nIndex + 1) == oldKids.size()) ) 
+    if ( !bIsPushedIn && ( (nIndex + 1) == static_cast<int>(oldKids.size())) ) 
     {
         for (std::vector<PdfObject*>::const_iterator itPages=vecPages.begin(); itPages!=vecPages.end(); ++itPages)
         {
